# openapi/loovie-server.openapi.yaml
openapi: 3.1.0
info:
  title: Loovie BYO Server Contract
  # Pre-1.0 while we are in public beta. The README and SECURITY.md
  # promise strict semver only at 1.0.0; the previous `1.0.0-beta.1`
  # value confused that promise. Bumps live in CHANGELOG.md and (post-
  # A.8) release-please will own this number.
  version: "0.1.0-beta.1"
  summary: The HTTP contract a third-party server implements to be drivable by the Loovie mobile app.
  description: |
    The HTTP API a "Bring Your Own" server must expose so the Loovie app
    can drive image and single-shot video generation on it. The Loovie
    backend NEVER calls this server; only the Loovie mobile app (on the
    user's device) does. The app stores the server URL and bearer token
    on the user's device only - they are needed there so the app can
    reach the server - and they are never sent to Loovie's backend
    servers or accessible to Loovie staff.

    Any HTTP server implementing this contract is valid - language and
    framework are your choice. We ship a ComfyUI reference implementation
    in `comfyui-loovie/` and a minimal FastAPI reference in
    `examples/minimal-server/`, but those are examples, not requirements.

    ## Auth

    A static bearer token chosen by the operator (`LOOVIE_API_TOKEN`).
    Required for all remote callers; localhost (`127.0.0.1`, `::1`) may
    bypass. If no token is configured the server MUST refuse remote
    requests (fail closed). The endpoints `/loovie/health` and
...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <LOOVIE_API_TOKEN>`. REQUIRED for remote
        callers. If no token is configured the server MUST refuse remote
        requests (fail closed). Localhost (127.0.0.1, ::1) bypasses.
  schemas:
    Envelope:
      type: object
      description: Standard response envelope used by every endpoint.
      required: [code]
      properties:
        code:
          type: integer
          description: HTTP-style status code mirrored into the body for clients that only read the body.
          examples: [200]
        msg:
          type: string
          description: Human-readable status or error message.
        data:
          type: object
          description: Endpoint-specific payload, present on success.
    ImageMode:
      type: string
      enum: [t2i, i2i]
      description: "`t2i` = text-to-image; `i2i` = image(+refs)-to-image."
    VideoMode:
      type: string
      enum: [t2v, i2v, fl2v]
      description: "`t2v` = text-to-video; `i2v` = image-to-video; `fl2v` = first+last-frame-to-video."
    Variant:
      type: string
      enum: [fast, pro]
      description: Quality/speed trade-off; `fast` for previews, `pro` for finals.
    ImageResolution:
      type: string
      enum: ["720p", "1K", "2K"]
    VideoResolution:
      type: string
      enum: ["720p", "1080p"]
    ImageAspectRatio:
      type: string
      enum: [auto, "1:1", "16:9", "9:16", "4:3"]
    VideoAspectRatio:
      type: string
      enum: [auto, "16:9", "9:16", "1:1"]
    Capabilities:
      type: object
      description: |
        What this server can do. The app probes this at save and at startup,
        caches the manifest per server, and uses it to gate the BYO picker
        tiers and constrain the BYO generation UI.
      required: [schemaVersion]
      properties:
        schemaVersion:
          type: integer
          const: 1
          description: Contract version. Bumps only when the manifest SHAPE changes.
        version:
          type: string
          description: Build semver of your implementation. Telemetry only.
        images:
          $ref: '#/components/schemas/ImageCapabilities'
        ss_videos:
          $ref: '#/components/schemas/VideoCapabilities'
    ImageCapabilities:
      type: object
      description: Image-generation capabilities. Omit this object entirely if the server has no image workflows.
      required: [capabilities, referenceCount, resolution, variants, aspectRatios]
      properties:
        capabilities:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ImageMode' }
        referenceCount:
          type: integer
          minimum: 0
          description: Max input reference images for the i2i mode.
        resolution:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ImageResolution' }
        variants:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Variant' }
        aspectRatios:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ImageAspectRatio' }
        promptCharLimit:
          type: integer
          minimum: 1
          default: 3000
          description: Max prompt length the server accepts. Optional; defaults to 3000 if absent.
    VideoCapabilities:
      type: object
      description: Single-shot video-generation capabilities. Omit this object entirely if the server has no video workflows.
      required: [capabilities, referenceCount, resolution, variants, aspectRatios, durations, supportsAudio]
      properties:
        capabilities:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/VideoMode' }
        referenceCount:
          type: integer
          minimum: 0
          description: Max input reference frames for i2v (fl2v always uses start+end).
        resolution:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/VideoResolution' }
        variants:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Variant' }
        aspectRatios:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/VideoAspectRatio' }
        durations:
          type: array
          minItems: 1
          items:
            type: integer
            minimum: 1
          description: Allowed clip durations in whole seconds (explicit list, may be non-contiguous).
        supportsAudio:
          type: boolean
          description: Whether the server can produce synced audio. When true, the app shows the "With Audio" toggle for the BYO video tier.
        promptCharLimit:
          type: integer
          minimum: 1
          default: 3000
...
    TaskCreated:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: object
              required: [taskId]
              properties:
                taskId:
                  type: string
                  examples: ["task_loovie_0a1b2c3d4e5f"]
                  description: Opaque task identifier. The app polls `/{kind}/status?taskId=...` with this.
    TaskState:
      type: string
      enum: [pending, processing, success, failed]
    TaskStatus:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: object
              required: [taskId, state]
              properties:
                taskId: { type: string }
                state: { $ref: '#/components/schemas/TaskState' }
                progress:
                  type: integer
                  minimum: 0
                  maximum: 100
                  description: 0-100. Optional for `pending`/`processing`; omitted on `success`/`failed`.
                elapsedSeconds: { type: number, description: "Seconds elapsed since task submission." }
                step: { type: integer, description: "Current workflow node index, if known." }
                totalSteps: { type: integer, description: "Total workflow node count, if known." }
                currentNode: { type: string, description: "Name of the currently executing node, if known." }
                resultJson:
                  type: string
                  description: |
                    Present only on `state: success`. A JSON-encoded string with shape
                    `{"resultUrls":["<url>"]}` where each URL is downloadable by the app
                    (no auth needed; URLs are signed or otherwise scoped by the server).
                totalSeconds: { type: number, description: "Total run time on `success`." }
                failCode:
                  type: string
                  description: |
                    Present only on `state: failed`. Short machine-readable code.
                    Common values: `WORKFLOW_NOT_FOUND`, `SUBMISSION_ERROR`, `EXECUTION_ERROR`, `OUTPUT_MISSING`, `HISTORY_MISSING`, `TIMEOUT`.
                    Servers MAY emit additional codes. The Loovie app namespaces non-prefixed codes as `BYO_<CODE>` at its boundary.
                failMsg:
                  type: string
                  description: 'Present only on `state: failed`. Human-readable failure description (will be shown to the operator, not the end user).'
    ImageCreateRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          maxLength: 3000
        mode: { $ref: '#/components/schemas/ImageMode' }
        variant: { $ref: '#/components/schemas/Variant' }
        aspect_ratio: { $ref: '#/components/schemas/ImageAspectRatio' }
        image_urls:
          type: array
          items: { type: string, format: uri }
          description: Reference images for i2i mode. Server downloads these.
        mask_url:
          type: string
          format: uri
          description: Optional inpainting mask URL.
        seed:
          type: integer
          default: -1
          description: Generation seed. `-1` means random.
        negative_prompt: { type: string }
        workflow:
          type: string
          description: Override workflow selection. Server resolves names like `flux-2-klein`.
        loras:
          type: array
          items: { type: object }
          description: Per-request LoRA stack (server-specific shape).
    VideoCreateRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          maxLength: 3000
        mode: { $ref: '#/components/schemas/VideoMode' }
        variant: { $ref: '#/components/schemas/Variant' }
        aspect_ratio: { $ref: '#/components/schemas/VideoAspectRatio' }
        resolution: { $ref: '#/components/schemas/VideoResolution' }
        duration:
          type: integer
          minimum: 1
          description: Clip duration in seconds. Must be one of the values in `ss_videos.durations`.
        start_frame_url:
          type: string
          format: uri
          description: Required for i2v and fl2v.
        end_frame_url:
          type: string
          format: uri
          description: Required for fl2v. Requires `start_frame_url`.
        with_audio:
          type: boolean
          description: 'Opt-in to synced audio. Only valid if `supportsAudio: true`.'
        seed:
          type: integer
          default: -1
    UploadJsonRequest:
      type: object
      required: [filename, data_base64]
      properties:
        filename: { type: string }
        data_base64: { type: string, description: "Base64-encoded file bytes." }
    UploadResponse:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                filename: { type: string }
                url:
                  type: string
                  description: A URL the server can use to resolve this upload in a subsequent generation request.
                sizeBytes: { type: integer }
paths:
...
  /loovie/capabilities:
    get:
      tags: [System]
      summary: What this server can do
      description: |
        UNAUTHENTICATED. The Loovie app probes this at save and at app
        start, caches the manifest per server, and uses it to gate the BYO
        picker. Servers omit the `images` or `ss_videos` section they
        can't serve, and the matching endpoints under `/images/*` or
        `/videos/*` MAY be omitted entirely.
      security: []
      operationId: getCapabilities
      responses:
        "200":
          description: Capability manifest.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Capabilities' }
              examples:
                imageAndVideo:
                  summary: Image + single-shot video
                  value:
                    schemaVersion: 1
                    version: "0.1.0"
                    images:
                      capabilities: ["t2i", "i2i"]
                      referenceCount: 4
                      resolution: ["720p", "1K", "2K"]
                      variants: ["fast", "pro"]
                      aspectRatios: ["auto", "1:1", "16:9", "9:16", "4:3"]
                      promptCharLimit: 3000
                    ss_videos:
                      capabilities: ["t2v", "i2v", "fl2v"]
                      referenceCount: 1
                      resolution: ["720p", "1080p"]
                      variants: ["fast", "pro"]
                      aspectRatios: ["auto", "16:9", "9:16", "1:1"]
                      durations: [1, 2, 3, 4, 5, 6, 7, 8]
                      supportsAudio: true
                      promptCharLimit: 3000
                imageOnly:
                  summary: Server with no video workflows installed
                  value:
                    schemaVersion: 1
                    version: "0.1.0"
                    images:
                      capabilities: ["t2i"]
                      referenceCount: 0
                      resolution: ["1K"]
                      variants: ["fast"]
                      aspectRatios: ["auto", "1:1"]
        "429":
          description: Rate-limited. Operators may throttle anonymous probes; the app retries with backoff.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /images/create:
    post:
      tags: [Image]
      summary: Start an image generation
      description: |
        Required only when `/loovie/capabilities` advertises an `images`
        section. Accepts a prompt and parameters; returns a `taskId` the
        caller polls via `/images/status`.
      operationId: createImage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImageCreateRequest' }
      responses:
        "200":
          description: Task accepted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskCreated' }
        "400":
          description: Malformed request (missing prompt, invalid enum, prompt > 3000 chars, unknown workflow, etc.).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        "422":
          description: Request body failed schema validation (e.g. wrong field types).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        "401":
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /images/status:
    get:
      tags: [Image]
      summary: Poll an image task
      description: Required only when `/loovie/capabilities` advertises an `images` section.
      operationId: getImageStatus
      parameters:
        - name: taskId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Current task state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskStatus' }
        "401":
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        "404":
          description: Task not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /videos/create:
    post:
      tags: [Video]
      summary: Start a single-shot video generation
      description: |
        Required only when `/loovie/capabilities` advertises an `ss_videos`
        section. `end_frame_url` requires `start_frame_url`. `with_audio`
        is only valid if `supportsAudio: true` in capabilities.
      operationId: createVideo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VideoCreateRequest' }
      responses:
        "200":
          description: Task accepted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskCreated' }
        "400":
          description: Malformed request.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        "422":
          description: Request body failed schema validation (e.g. wrong field types).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        "401":
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /videos/status:
    get:
      tags: [Video]
      summary: Poll a video task
      description: Required only when `/loovie/capabilities` advertises an `ss_videos` section.
      operationId: getVideoStatus
      parameters:
        - name: taskId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Current task state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskStatus' }
        "401":
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        "404":
          description: Task not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
  /loovie/upload:
    post:
      tags: [Upload]
      summary: Stage an input file
      description: |
        Optional operator-side helper. The Loovie app uploads reference
        images using whatever URL scheme the server-side `image_urls`
        accepts (typically signed URLs the Loovie app already has). This
        endpoint exists for operator-side tooling that wants to push
        files into the server's input store and reference them in a
        subsequent generation.
      operationId: upload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
          application/json:
            schema: { $ref: '#/components/schemas/UploadJsonRequest' }
      responses:
        "200":
          description: Stored.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadResponse' }
        "400":
          description: Malformed input (invalid base64, non-ASCII filename, missing required fields).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        "422":
          description: Request body failed schema validation (e.g. wrong field types).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        "401":
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }